Updates made to resolve Xcode version error on app submittal.

+++ SpeedyScan - Document Scanner/Controller/CaptureDetailVC.swift
@@ -396,7 +396,7 @@ class CaptureDetailVC: UIViewController {
 	private func compressImage(image: UIImage) -> UIImage {
 		//Downsampling of image.
 		
-		let imageReductionFactor = 1.0
+		let imageReductionFactor = CGFloat(1.0)
 		
 		let reducedImageSize = CGSize(width: image.size.width * imageReductionFactor, height: image.size.height * imageReductionFactor)
